+++ noodles-bam/src/record/codec.rs
@@ -1,6 +1,7 @@
-#![doc(hidden)]
+//! BAM record codec for encoding and decoding.
 
 pub mod decoder;
 pub mod encoder;
 
-pub(crate) use self::{decoder::decode, encoder::encode};
+pub use self::decoder::{DecodeError, decode};
+pub use self::encoder::encode;